+++ src/database/postgres.rs
@@ -281,68 +281,18 @@ impl Queries for PostgresClient {
         Ok(row.get(0))
     }
 
-    async fn top_record(&self, map_uid: &str) -> Result<Option<RecordDetailed>> {
-        let conn = self.0.get().await?;
-        let stmt = r#"
-            SELECT p.login, p.nick_name, r.millis, r.timestamp, s.cp_millis
-            FROM (
-                SELECT player_login, millis, timestamp
-                FROM steward.record
-                WHERE map_uid = $1
-                ORDER BY millis ASC
-                LIMIT 1
-            ) r
-            RIGHT JOIN steward.sector s ON r.player_login = s.player_login
-            LEFT JOIN steward.player p ON r.player_login = p.login
-            WHERE r.player_login IS NOT NULL AND s.map_uid = $1
-            ORDER BY s.index ASC
-        "#;
-        let rows = conn.query(stmt, &[&map_uid]).await?;
-        Ok(rows.first().map(|row| RecordDetailed {
-            map_rank: 1,
-            player_login: row.get("login"),
-            player_nick_name: GameString::from(row.get("nick_name")),
-            timestamp: row.get("timestamp"),
-            millis: row.get("millis"),
-            cp_millis: rows.iter().map(|row| row.get("cp_millis")).collect(),
-        }))
-    }
-
-    async fn top_records(&self, map_uid: &str, limit: i64) -> Result<Vec<Record>> {
-        let conn = self.0.get().await?;
-        let stmt = r#"
-            SELECT p.login, p.nick_name, r.timestamp, r.millis
-            FROM steward.record r
-            LEFT JOIN steward.player p
-            ON r.player_login = p.login
-            WHERE map_uid = $1
-            ORDER BY r.millis ASC
-            LIMIT $2
-        "#;
-        let rows = conn.query(stmt, &[&map_uid, &limit]).await?;
-
-        Ok(rows
-            .iter()
-            .map(|row| Record {
-                player_login: row.get("login"),
-                player_nick_name: GameString::from(row.get("nick_name")),
-                timestamp: row.get("timestamp"),
-                millis: row.get("millis"),
-            })
-            .collect())
-    }
-
-    async fn player_records(
+    async fn records(
         &self,
-        map_uid: &str,
+        map_uids: Vec<&str>,
         player_logins: Vec<&str>,
-    ) -> Result<Vec<RecordDetailed>> {
+        limit_per_map: Option<i64>,
+    ) -> Result<Vec<Record>> {
         let conn = self.0.get().await?;
         let stmt = r#"
             SELECT
-                r.pos, r.millis, r.timestamp,
+                r.map_uid, r.pos, r.millis, r.timestamp,
                 p.login, p.nick_name,
-                s.cp_millis
+                s.cp_millis, s.cp_speed
             FROM (
                 SELECT
                    map_uid,
@@ -353,31 +303,51 @@ impl Queries for PostgresClient {
                       ORDER BY millis ASC
                    ) pos
                 FROM steward.record
-                WHERE map_uid = $1
+                WHERE CARDINALITY($1::text[]) = 0 OR map_uid = ANY($1::text[])
+                LIMIT $3
             ) r
             NATURAL JOIN (
                 SELECT
                     map_uid,
                     player_login,
-                    ARRAY_AGG(cp_millis ORDER BY index ASC) cp_millis
+                    ARRAY_AGG(cp_millis ORDER BY index ASC) cp_millis,
+                    ARRAY_AGG(cp_speed ORDER BY index ASC) cp_speed
                 FROM steward.sector
-                WHERE map_uid = $1 AND player_login = ANY($2)
+                WHERE CARDINALITY($2::text[]) = 0 OR player_login = ANY($2::text[])
                 GROUP BY map_uid, player_login
             ) s
             INNER JOIN steward.player p ON r.player_login = p.login
         "#;
-        let rows = conn.query(stmt, &[&map_uid, &player_logins]).await?;
-        Ok(rows
+        let rows = conn
+            .query(stmt, &[&map_uids, &player_logins, &limit_per_map])
+            .await?;
+        let records = rows
             .into_iter()
-            .map(|row| RecordDetailed {
-                map_rank: row.get("pos"),
-                player_login: row.get("login"),
-                player_nick_name: GameString::from(row.get("nick_name")),
-                timestamp: row.get("timestamp"),
-                millis: row.get("millis"),
-                cp_millis: row.get::<_, Vec<i32>>("cp_millis"),
+            .map(|row| {
+                let cp_millis: Vec<i32> = row.get("cp_millis");
+                let cp_speed: Vec<f32> = row.get("cp_speed");
+                let sectors = cp_millis
+                    .into_iter()
+                    .zip(cp_speed.into_iter())
+                    .enumerate()
+                    .map(|(idx, (millis, speed))| RecordSector {
+                        index: idx as i32,
+                        cp_millis: millis,
+                        cp_speed: speed,
+                    })
+                    .collect();
+                Record {
+                    map_uid: row.get("map_uid"),
+                    map_rank: row.get("pos"),
+                    player_login: row.get("login"),
+                    player_nick_name: GameString::from(row.get("nick_name")),
+                    timestamp: row.get("timestamp"),
+                    millis: row.get("millis"),
+                    sectors,
+                }
             })
-            .collect())
+            .collect();
+        Ok(records)
     }
 
     async fn nb_players_with_record(&self) -> Result<i64> {
@@ -435,28 +405,19 @@ impl Queries for PostgresClient {
 
         let insert_record_stmt = r#"
             INSERT INTO steward.record
-                (player_login, map_uid, millis, validation, ghost, timestamp)
+                (player_login, map_uid, millis, timestamp)
             VALUES
-                ($1, $2, $3, $4, $5, $6)
+                ($1, $2, $3, $4)
             ON CONFLICT (player_login, map_uid)
             DO UPDATE SET
                 millis = excluded.millis,
-                validation = excluded.validation,
-                ghost = excluded.ghost,
                 timestamp = excluded.timestamp
         "#;
 
         let _ = transaction
             .execute(
                 insert_record_stmt,
-                &[
-                    &rec.player_login,
-                    &rec.map_uid,
-                    &rec.millis,
-                    &rec.validation,
-                    &rec.ghost,
-                    &rec.timestamp,
-                ],
+                &[&rec.player_login, &rec.map_uid, &rec.millis, &rec.timestamp],
             )
             .await?;
 
@@ -605,29 +566,6 @@ impl Queries for PostgresClient {
 
         transaction.commit().await?;
         Ok(maybe_map)
-    }
-
-    async fn delete_old_ghosts(&self, max_rank: i64) -> Result<u64> {
-        assert!(max_rank > 0, "tried to delete every ghost replay");
-
-        let conn = self.0.get().await?;
-        let stmt = r#"
-            DELETE FROM steward.record a
-            USING (
-                SELECT
-                   player_login,
-                   map_uid,
-                   RANK () OVER (
-                      PARTITION BY player_login, map_uid
-                      ORDER BY millis ASC
-                   ) pos
-                FROM steward.record
-            ) b
-            WHERE a.player_login = b.player_login AND a.map_uid = b.map_uid AND b.pos > $1
-        "#;
-        let nb_deleted = conn.execute(stmt, &[&max_rank]).await?;
-
-        Ok(nb_deleted)
     }
 }