+++ tests/integration/main.rs
@@ -29,7 +29,8 @@ use steward::server::{GameString, PlayerInfo};
 // [ ] top_record
 // [ ] top_records
 // [ ] player_record
-// [x] player_records
+// [ ] records
+//     - test with limit
 // [ ] nb_players_with_record
 // [x] maps_without_player_record
 // [x] record_preview
@@ -40,7 +41,6 @@ use steward::server::{GameString, PlayerInfo};
 // [ ] map_rankings
 // [ ] delete_player
 // [ ] delete_map
-// [ ] delete_old_ghosts
 
 /// Spins up a Postgres database in a Docker container.
 async fn clean_db() -> Result<Arc<dyn Database>> {
@@ -197,7 +197,7 @@ async fn test_player_record_some() -> Result<()> {
     db.upsert_map(&map).await?;
     db.upsert_record(&rec).await?;
 
-    let expected = record_detailed(1, "nickname", rec);
+    let expected = record(1, "nickname", rec);
     let expected = Some(expected);
 
     let actual = db.player_record("uid1", "login").await?;
@@ -207,7 +207,7 @@ async fn test_player_record_some() -> Result<()> {
 }
 
 #[tokio::test]
-async fn test_player_records_single() -> Result<()> {
+async fn test_records_single() -> Result<()> {
     let db = clean_db().await?;
 
     let player = player_info("login", "nickname");
@@ -217,17 +217,17 @@ async fn test_player_records_single() -> Result<()> {
     db.upsert_map(&map).await?;
     db.upsert_record(&rec).await?;
 
-    let expected = record_detailed(1, "nickname", rec);
+    let expected = record(1, "nickname", rec);
     let expected = vec![expected];
 
-    let actual = db.player_records("uid1", vec!["login"]).await?;
+    let actual = db.records(vec!["uid1"], vec!["login"], None).await?;
     assert_eq!(expected, actual);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_player_records_multiple_players() -> Result<()> {
+async fn test_records_multiple_players() -> Result<()> {
     let db = clean_db().await?;
 
     let player1 = player_info("login1", "nickname1");
@@ -241,18 +241,20 @@ async fn test_player_records_multiple_players() -> Result<()> {
     db.upsert_record(&rec1).await?;
     db.upsert_record(&rec2).await?;
 
-    let expected1 = record_detailed(1, "nickname1", rec1);
-    let expected2 = record_detailed(2, "nickname2", rec2);
+    let expected1 = record(1, "nickname1", rec1);
+    let expected2 = record(2, "nickname2", rec2);
     let expected = vec![expected1, expected2];
 
-    let actual = db.player_records("uid1", vec!["login1", "login2"]).await?;
+    let actual = db
+        .records(vec!["uid1"], vec!["login1", "login2"], None)
+        .await?;
     assert_eq!(expected, actual);
 
     Ok(())
 }
 
 #[tokio::test]
-async fn test_player_records_multiple_maps() -> Result<()> {
+async fn test_records_multiple_maps() -> Result<()> {
     let db = clean_db().await?;
 
     let player1 = player_info("login1", "nickname1");
@@ -268,10 +270,12 @@ async fn test_player_records_multiple_maps() -> Result<()> {
     db.upsert_record(&rec1).await?;
     db.upsert_record(&rec2).await?;
 
-    let expected = record_detailed(1, "nickname1", rec1);
+    let expected = record(1, "nickname1", rec1);
     let expected = vec![expected];
 
-    let actual = db.player_records("uid1", vec!["login1", "login2"]).await?;
+    let actual = db
+        .records(vec!["uid1"], vec!["login1", "login2"], None)
+        .await?;
     assert_eq!(expected, actual);
 
     Ok(())
@@ -574,8 +578,6 @@ fn record_evidence(login: &str, map_uid: &str, millis: i32) -> RecordEvidence {
         map_uid: map_uid.to_string(),
         millis,
         timestamp: now(),
-        validation: "validation replay".as_bytes().to_owned(),
-        ghost: Some("ghost replay".as_bytes().to_owned()),
         sectors: (0..5)
             .map(|i| RecordSector {
                 index: i,
@@ -586,14 +588,15 @@ fn record_evidence(login: &str, map_uid: &str, millis: i32) -> RecordEvidence {
     }
 }
 
-fn record_detailed(pos: i64, nick_name: &str, ev: RecordEvidence) -> RecordDetailed {
-    RecordDetailed {
+fn record(pos: i64, nick_name: &str, ev: RecordEvidence) -> Record {
+    Record {
+        map_uid: ev.map_uid,
         map_rank: pos,
         player_login: ev.player_login,
         player_nick_name: GameString::from(nick_name.to_string()),
         millis: ev.millis,
         timestamp: ev.timestamp,
-        cp_millis: ev.sectors.iter().map(|sector| sector.cp_millis).collect(),
+        sectors: ev.sectors,
     }
 }